refactor(agent): split agent.rs into per-concern submodules
The 1414-line agent.rs bundled six distinct concerns into one file: slash
parsing, LLM JSON extraction, skill dispatch, plain chat, upload decode,
and SSE encoding. Split into:

  agent.rs       — entry: doc, types, impl Block, /commands, /chat dispatcher
  agent/slash.rs — parse_slash, lookup_skill_tool, ParamExtraction,
                   is_prompt_shaped, build_skill_params, extract_via_llm,
                   strip_to_json, truncate, collect_chat_text
  agent/dispatch.rs — run_skill_dispatch, ToolOutcome, parse_skill_response
  agent/chat.rs  — run_plain_chat
  agent/messages.rs — openai_json_to_chat_message
  agent/uploads.rs  — decode_uploads, build_upload_history_prefix
  agent/sse.rs   — encode_sse_event, sse_response, error_response

No behavior change. Each submodule's tests moved alongside its functions.
Visibility: items shared between submodules use pub(super); items private
to a single concern stay default-private; AgentError / UploadEntry /
SKILL_PREFIX stay private to agent.rs (descendants inherit access).

Largest file shrank from 1414 to 404 lines. Net LOC change: +25
(per-file imports + module declarations), traded for navigability.

Verified: 32/32 unit tests pass; cargo check clean on host and
wasm32-unknown-unknown.
